cluster: render tispark configs and scripts
